refactoring into smaller components, improved UI
